requirements.txt
asgiref==3.4.1
certifi==2021.10.8
distlib==0.3.4
Django==3.2.12
django-cors-headers==3.10.1
django-environ==0.8.1
djangorestframework==3.13.1
filelock==3.4.1
gunicorn==20.1.0
pip==21.2.4
platformdirs==2.4.0
PyMySQL==1.0.2
pytz==2021.3
setuptools==58.1.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
virtualenv==20.13.0
wheel==0.37.1